update MSToolkit to commit 2021e7 (12/3/19)
git-svn-id: https://svn.code.sf.net/p/comet-ms/code@1374 299d2d02-8f3b-485e-a07e-71e8e5e16577

+++ branches/release_2019013/MSToolkit/include/CSearchDatabase.h
@@ -32,8 +32,8 @@ public:
   std::string id;
   std::string location;
   std::string name;
-  long numDatabaseSequences;
-  long numResidues;
+  int numDatabaseSequences;
+  int numResidues;
   std::string releaseDate;
   std::string version;
   sExternalFormatDocumentation externalFormatDocumentation;
